# Optimize int8 layers in DNN modules by using RISC-V Vector intrinsic.