@mshabunin 2024-03-19
Which compiler did you use? Currently OpenCV uses v0.10 of RVV intrinsics + compatibility layers to v0.11 and v0.12. It seems you've used the latest intrinsics version. Perhaps these code blocks should be guarded: `#if defined(__riscv_v_intrinsic) &&  __riscv_v_intrinsic>=12000`. Alternatively we can consider upgrading all our intrinsics to the latest version and drop support of older compilers (intrinsics version check would be added to the current compile check).