@hanliutong 2024-03-20
I'm using clang 16.0.6. These code can work when RVV intrinsic version is 0.11 or 0.12, so both clang 16.x and 17.x (and trunk) should work. But with clang 15.x, which only supports version 0.10, this code will not compile.

Considering that v1.0-rc1 of RVV intrinsic has been released and looks like it is very close to stable, I think it is reasonable to upgrade all our intrinsics to the v1.0 (when it is officially released). 

And for now, we may use ` #if defined(__riscv_v_intrinsic) && __riscv_v_intrinsic>=11000` to temporary guard there code.