@hanliutong 2024-03-20
Compiler compatibility: https://godbolt.org/z/ns9afhTae