@mshabunin 2024-03-28
For some reason some tests fail when built with GCC 13.2+ (qemu 8.2.1):
```
qemu-riscv64 \
	-L ${sysroot} \
	-cpu rv64,v=true,vext_spec=v1.0 \
./bin/opencv_test_$t --gtest_filter="*Int8_layers*"
...
[  FAILED  ] 11 tests, listed below:
[  FAILED  ] Test_Int8_layers.Convolution2D/0, where GetParam() = OCV/CPU
[  FAILED  ] Test_Int8_layers.Padding/0, where GetParam() = OCV/CPU
[  FAILED  ] Test_Int8_layers.AvePooling/0, where GetParam() = OCV/CPU
[  FAILED  ] Test_Int8_layers.MaxPooling/0, where GetParam() = OCV/CPU
[  FAILED  ] Test_Int8_layers.Softmax_slim_TF/0, where GetParam() = OCV/CPU
[  FAILED  ] Test_Int8_layers.Concat/0, where GetParam() = OCV/CPU
[  FAILED  ] Test_Int8_layers.InnerProduct/0, where GetParam() = OCV/CPU
[  FAILED  ] Test_Int8_layers.Reshape/0, where GetParam() = OCV/CPU
[  FAILED  ] Test_Int8_layers.Slice_4d_tf/0, where GetParam() = OCV/CPU
[  FAILED  ] Test_Int8_layers.Slice_strided_tf/0, where GetParam() = OCV/CPU
[  FAILED  ] Test_Int8_layers.Eltwise/0, where GetParam() = OCV/CPU
```
Examples of failures:
```
[ RUN      ] Test_Int8_layers.Convolution2D/0, where GetParam() = OCV/CPU
/work/opencv/modules/dnn/test/test_common.impl.hpp:76: Failure
Expected: (normL1) <= (l1), actual: 0.49239 vs 0.00413
single_conv  |ref| = 4.2324190139770508
/work/opencv/modules/dnn/test/test_common.impl.hpp:79: Failure
Expected: (normInf) <= (lInf), actual: 3.21996 vs 0.02201
single_conv  |ref| = 4.2324190139770508
/work/opencv/modules/dnn/test/test_common.impl.hpp:76: Failure
Expected: (normL1) <= (l1), actual: 1.6915 vs 0.0193
atrous_conv2d_valid  |ref| = 9.5424537658691406
```

With clang 17, these tests pass. I'm not sure whether the problem is with GCC or OpenCV. Perhaps we can merge it as-is and try to find the problem later.