@hanliutong 2024-03-28
I can also reproduce those tests fail when I use gcc 14.0.1(trunk). Working on it.