@hanliutong 2024-03-29
Fixed. Test are passed both with clang and gcc now. Thanks for your tests on gcc @mshabunin !

GCC is right, test passed on clang is lucky. When accumulating vectors, we should use "tail undisturbed" (tu) to make sure that the elements after `vl` not change (keeping the last accumulation result). I incorrectly used "tail agnostic" (ta) earlier.